Move race combinator into subdirectory
