+++ packages/core/src/index.ts
@@ -8,13 +8,13 @@ export { Operation, Resource } from './operation';
 export { Effection } from './effection';
 export { deprecated } from './deprecated';
 export { Deferred } from './deferred';
-export { race } from './race';
 
 export { sleep } from './operations/sleep';
 export { ensure } from './operations/ensure';
 export { timeout } from './operations/timeout';
 export { withTimeout } from './operations/with-timeout';
 export { spawn } from './operations/spawn';
+export { race } from './operations/race';
 
 export function run<TOut>(operation?: Operation<TOut>, options?: TaskOptions): Task<TOut> {
   return Effection.root.spawn(operation, options);
